requirements.txt
jc==1.22.4
gunicorn>=20.0.4
click==7.1.2
Flask==1.1.2
Flask-WTF==0.14.3
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
Pygments==2.6.1
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
Werkzeug==1.0.1
WTForms==2.3.3
xmltodict==0.12.0
mistune==2.0.2
